requirements_.txt
tflite-runtime==2.14.0
numpy==1.26.4
pillow==10.3.0
streamlit==1.35.0
